+++ src/main/java/com/smart/tailor/utils/request/MaterialRequest.java
@@ -31,7 +31,7 @@ public class MaterialRequest {
 
     @NotNull(message = "basePrice is required")
     @Min(value = 0, message = "basePrice can not less than 0")
-    private Double basePrice;
+    private Integer basePrice;
 
     @Override
     public String toString() {
